msm8937_64: Enable channel bonding for all channels in world reg domain
If country in NV is 00, driver indicates it as CUSTOM REGULATORY
to kernel during registration. In case of CUSTOM REGULATORY kernel
doesn't call reg notifier to update the channel list in driver
from db.txt. Thus the driver channel list is prepared from NV.
Now in NV the upper bonding channels are disabled for world reg
domain and thus driver doesn't allow any of these channels in
40/80 Mhz mode.

As part of the fix enable channel bonding for all channels in world
reg domain

Change-Id: I9dcdac76026de7adf5ab8baa13b2f89ed0bd011d
CRs-Fixed: 956141